the v19.9 should be probably obsolete with the polyfill
